refactor(services): PR-3.8 model_downloader.py 1068 → models/ 4-way 拆
services/models_downloader.py 1068 行单文件按职责切 4 文件，shim 留原 import
路径透明转发，业务调用方零改动。

  paths.py 293     模型路径常量 + 本地 target 解析 + safe_dir_name
  sources.py 218   下载源 / endpoint / 镜像 / download_flat[_ms] 原语
  downloader.py 431  per-model 高层下载 + 异步 DownloadStatus
  catalog.py 226   build_catalog 扫盘汇总

shim：services/model_downloader.py 用 sys.modules[__name__] = _models 把整个
旧模块对象别名到新包，旧 `from studio.services.model_downloader import X`
经包 __init__ re-export 仍工作。

跨边界调用走属性访问：downloader.py 内调用 sources 函数必须用
`_sources.download_flat(...)` 而非 `from .sources import download_flat`
+ 直接调用——后者会 bind 成 module-local name，让测试
`monkeypatch.setattr("studio.services.models.sources.download_flat", ...)`
对 downloader 内的调用无效。__init__.py 顶部注释 + downloader.py 顶部都
留警示。

调用方修：services/presets/__init__.py / tagging/cltagger.py / tagging/wd14.py /
data_io/train_io.py 4 处 `from ..models import downloader as model_downloader`
原意是「拿整个 model_downloader 命名空间」，PR-3 services 拆包后 `models`
变包名，写法误把 downloader 子模块当包用，paths/catalog 的函数（含
default_paths_for_new_version / cltagger_target_root / wd14_target_dir /
models_root）就找不到了。改为 `from .. import models as model_downloader`
继续用包对象作命名空间。

onnx_base.py 的 safe_dir_name 顺手搬到 paths.py，原位置留 noqa F401 shim
打破潜在循环 (models/paths ← tagging/onnx_base ← models/downloader)。
