+++ studio/services/models/__init__.py
@@ -1,0 +1,124 @@
+"""models 子包 —— PR-3.8 把 model_downloader.py 1068 行 4-way 拆。
+
+  paths.py       常量 + 本地路径解析
+  sources.py     下载源 / endpoint / 镜像 / download_flat 原语
+  downloader.py  per-model 高层下载 + async 状态
+  catalog.py     build_catalog 扫盘汇总
+
+re-export 完整公共名供 `from studio.services.models import X` 使用。
+原 `studio.services.model_downloader` 通过 shim 文件 sys.modules 别名到 **本包**
+（不是 downloader 子模块），原 `from studio.services.model_downloader import X` 仍工作。
+"""
+from .catalog import build_catalog
+from .downloader import (
+    DownloadStatus,
+    _DOWNLOADS,
+    _LOCK,
+    download_anima_main,
+    download_anima_vae,
+    download_cltagger,
+    download_qwen3,
+    download_t5_tokenizer,
+    download_taeflux,
+    download_upscaler,
+    download_upscaler_custom,
+    download_wd14,
+    get_status_snapshot,
+    start_download_async,
+    trigger,
+)
+from .paths import (
+    ANIMA_REPO,
+    ANIMA_VAE_PATH,
+    ANIMA_VARIANTS,
+    CLTAGGER_VERSIONS,
+    DEFAULT_UPSCALER,
+    LATEST_ANIMA,
+    QWEN_FILES,
+    QWEN_REPO,
+    T5_FILES,
+    T5_REPO,
+    TAEFLUX_FILES,
+    TAEFLUX_REPO,
+    UPSCALER_EXTS,
+    UPSCALER_VARIANTS,
+    WD14_FILES,
+    anima_main_target,
+    anima_vae_target,
+    cltagger_target_root,
+    default_paths_for_new_version,
+    find_anima_main,
+    find_upscaler,
+    models_root,
+    qwen_dir,
+    selected_anima_variant,
+    selected_upscaler,
+    t5_tokenizer_dir,
+    taeflux_available,
+    taeflux_dir,
+    upscaler_dir,
+    upscaler_target,
+    wd14_target_dir,
+)
+from .sources import (
+    MS_ANIMA_TEXT_ENCODER_PATH,
+    _get_download_source,
+    _ms_token,
+    _ms_wd14_repo_id,
+    _resolve_endpoint,
+    download_flat,
+    download_flat_ms,
+    setup_mirror,
+)
+
+__all__ = [
+    "ANIMA_REPO",
+    "ANIMA_VAE_PATH",
+    "ANIMA_VARIANTS",
+    "CLTAGGER_VERSIONS",
+    "DEFAULT_UPSCALER",
+    "DownloadStatus",
+    "LATEST_ANIMA",
+    "MS_ANIMA_TEXT_ENCODER_PATH",
+    "QWEN_FILES",
+    "QWEN_REPO",
+    "T5_FILES",
+    "T5_REPO",
+    "TAEFLUX_FILES",
+    "TAEFLUX_REPO",
+    "UPSCALER_EXTS",
+    "UPSCALER_VARIANTS",
+    "WD14_FILES",
+    "anima_main_target",
+    "anima_vae_target",
+    "build_catalog",
+    "cltagger_target_root",
+    "default_paths_for_new_version",
+    "download_anima_main",
+    "download_anima_vae",
+    "download_cltagger",
+    "download_flat",
+    "download_flat_ms",
+    "download_qwen3",
+    "download_t5_tokenizer",
+    "download_taeflux",
+    "download_upscaler",
+    "download_upscaler_custom",
+    "download_wd14",
+    "find_anima_main",
+    "find_upscaler",
+    "get_status_snapshot",
+    "models_root",
+    "qwen_dir",
+    "selected_anima_variant",
+    "selected_upscaler",
+    "setup_mirror",
+    "start_download_async",
+    "t5_tokenizer_dir",
+    "taeflux_available",
+    "taeflux_dir",
+    "trigger",
+    "upscaler_dir",
+    "upscaler_target",
+    "wd14_target_dir",
+]
