+++ studio/services/tagging/cltagger.py
@@ -14,7 +14,7 @@ import numpy as np
 from PIL import Image, ImageOps
 
 from ... import secrets
-from ..models import downloader as model_downloader
+from .. import models as model_downloader
 from .onnx_base import OnnxTaggerBase
 
 logger = logging.getLogger(__name__)
